requirements.txt
black==23.12.1
click==8.1.7
dnspython==2.4.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
python-dotenv==1.0.0
black==23.12.1
click==8.1.7
dnspython==2.4.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
python-dotenv==1.0.0
black==23.12.1
click==8.1.7
dnspython==2.4.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
python-dotenv==1.0.0
black==23.12.1
click==8.1.7
dnspython==2.4.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pymongo==4.6.1
python-dotenv==1.0.0
